CDD-2180: abstracted out show more logic and added tests
